deleted one of the two copies of adyen-checkout
